feat: bundle JavaScript with rollup
