+++ main.js
@@ -1,146 +1,2 @@
-import { setupServiceWorker } from './js/sw-register.js';
-import { inicialitza } from './js/init.js';
-import { setupRouter } from './js/router.js';
-import { closeChart, handleResize } from './js/graficos.js';
-
-setupServiceWorker();
-
-
-document.querySelectorAll('.menu-toggle').forEach(btn => {
-  btn.addEventListener('click', () => {
-    const target = document.getElementById(btn.dataset.target);
-    const isVisible = target.style.display === 'flex';
-    document.querySelectorAll('.submenu').forEach(sm => {
-      sm.style.display = 'none';
-      sm.setAttribute('aria-hidden', 'true');
-    });
-    document.querySelectorAll('.menu-toggle').forEach(b => b.setAttribute('aria-expanded', 'false'));
-    if (!isVisible) {
-      target.style.display = 'flex';
-      target.setAttribute('aria-hidden', 'false');
-      btn.setAttribute('aria-expanded', 'true');
-      const firstButton = target.querySelector('button');
-      if (firstButton) firstButton.focus();
-    } else {
-      target.style.display = 'none';
-      target.setAttribute('aria-hidden', 'true');
-      btn.setAttribute('aria-expanded', 'false');
-    }
-    document.getElementById('filters-row').style.display = 'none';
-    document.getElementById('classificacio-filters').style.display = 'none';
-    document.getElementById('torneig-buttons').style.display = 'none';
-    document.getElementById('torneig-title').style.display = 'none';
-    document.getElementById('torneig-category-buttons').style.display = 'none';
-    document.getElementById('continu3b-buttons').style.display = 'none';
-    document.getElementById('content').style.display = 'none';
-  });
-});
-
-document.getElementById('btn-ranking').addEventListener('click', () => {
-  document.getElementById('filters-row').style.display = 'flex';
-  document.getElementById('classificacio-filters').style.display = 'none';
-  document.getElementById('torneig-buttons').style.display = 'none';
-  document.getElementById('torneig-title').style.display = 'none';
-  document.getElementById('torneig-category-buttons').style.display = 'none';
-  document.getElementById('continu3b-buttons').style.display = 'none';
-  document.getElementById('content').style.display = 'block';
-  mostraRanquing();
-});
-
-document.getElementById('btn-classificacio').addEventListener('click', () => {
-  document.getElementById('filters-row').style.display = 'none';
-  document.getElementById('classificacio-filters').style.display = 'flex';
-  document.getElementById('torneig-buttons').style.display = 'none';
-  document.getElementById('torneig-title').style.display = 'none';
-  document.getElementById('torneig-category-buttons').style.display = 'none';
-  document.getElementById('continu3b-buttons').style.display = 'none';
-  document.getElementById('content').style.display = 'block';
-  mostraClassificacio();
-});
-
-document.getElementById('btn-agenda').addEventListener('click', () => {
-  document.getElementById('filters-row').style.display = 'none';
-  document.getElementById('classificacio-filters').style.display = 'none';
-  document.getElementById('torneig-buttons').style.display = 'none';
-  document.getElementById('torneig-title').style.display = 'none';
-  document.getElementById('torneig-category-buttons').style.display = 'none';
-  document.getElementById('continu3b-buttons').style.display = 'none';
-  document.getElementById('content').style.display = 'block';
-  mostraAgenda();
-});
-
-document.getElementById('btn-horari').addEventListener('click', () => {
-  document.getElementById('filters-row').style.display = 'none';
-  document.getElementById('classificacio-filters').style.display = 'none';
-  document.getElementById('torneig-buttons').style.display = 'none';
-  document.getElementById('torneig-title').style.display = 'none';
-  document.getElementById('torneig-category-buttons').style.display = 'none';
-  document.getElementById('continu3b-buttons').style.display = 'none';
-  document.getElementById('content').style.display = 'block';
-  mostraHorari();
-});
-
-document.getElementById('btn-enllacos').addEventListener('click', () => {
-  document.getElementById('filters-row').style.display = 'none';
-  document.getElementById('classificacio-filters').style.display = 'none';
-  document.getElementById('torneig-buttons').style.display = 'none';
-  document.getElementById('torneig-title').style.display = 'none';
-  document.getElementById('torneig-category-buttons').style.display = 'none';
-  document.getElementById('continu3b-buttons').style.display = 'none';
-  document.getElementById('content').style.display = 'block';
-  mostraEnllacos();
-});
-
-document.getElementById('btn-continu3b').addEventListener('click', () => {
-  document.getElementById('filters-row').style.display = 'none';
-  document.getElementById('classificacio-filters').style.display = 'none';
-  document.getElementById('torneig-buttons').style.display = 'none';
-  document.getElementById('torneig-title').style.display = 'none';
-  document.getElementById('torneig-category-buttons').style.display = 'none';
-  const cont = document.getElementById('content');
-  cont.style.display = 'block';
-  const btns = document.getElementById('continu3b-buttons');
-  btns.style.display = 'flex';
-  mostraContinu3B();
-});
-
-document.getElementById('btn-torneig').addEventListener('click', () => {
-  document.getElementById('filters-row').style.display = 'none';
-  document.getElementById('classificacio-filters').style.display = 'none';
-  document.getElementById('torneig-buttons').style.display = 'flex';
-  document.getElementById('torneig-category-buttons').style.display = 'none';
-  document.getElementById('continu3b-buttons').style.display = 'none';
-  const cont = document.getElementById('content');
-  cont.style.display = 'block';
-  cont.innerHTML = '';
-  fetch('data/modalitat.json')
-    .then(r => r.json())
-    .then(d => {
-      const modalObj = Array.isArray(d) ? (d[0] || {}) : (d || {});
-      const title = document.getElementById('torneig-title');
-      title.textContent = `Social Modalitat ${Array.isArray(d) ? d.map(m => m.Modalitat).join(', ') : (d.Modalitat || '')}`;
-      title.style.display = 'block';
-    })
-    .catch(() => {
-      const title = document.getElementById('torneig-title');
-      title.style.display = 'none';
-    });
-});
-
-document.querySelectorAll('#torneig-buttons button').forEach(btn => {
-  btn.addEventListener('click', () => {
-    const file = btn.dataset.file;
-    fetch(`data/${file}`)
-      .then(r => r.json())
-      .then(d => mostraTorneig(d, file))
-      .catch(() => {
-        document.getElementById('content').innerHTML = '<p>Error carregant dades.</p>';
-      });
-  });
-});
-
-
-document.getElementById('close-chart').addEventListener('click', closeChart);
-window.addEventListener('resize', handleResize);
-
-inicialitza();
+let e=null;function t(t,n){const o=["3 BANDES","BANDA","LLIURE"],a=o.map(e=>l.filter(n=>n.Jugador===t&&n.Modalitat===e).map(e=>({any:parseInt(e.Any,10),mitjana:parseFloat(e.Mitjana)})).sort((e,t)=>e.any-t.any)),s=[...new Set(a.flat().map(e=>e.any))].sort((e,t)=>e-t),d=["blue","green","red"],i=o.map((e,t)=>({label:e,data:s.map(e=>{const n=a[t].find(t=>t.any===e);return n?Number.parseFloat(n.mitjana):null}),borderColor:d[t],backgroundColor:d[t],tension:.2,fill:!1})),r=document.getElementById("chart-canvas"),y=document.getElementById("chart-overlay");y&&(y.style.display="flex",e=document.activeElement);const m=document.getElementById("close-chart");m&&m.focus(),E();const u=document.getElementById("chart-title");u&&(u.textContent=n);const p=r.getContext("2d");c&&c.destroy(),g(new Chart(p,{type:"line",data:{labels:s,datasets:i},options:{plugins:{legend:{position:"right"}},scales:{x:{title:{display:!0,text:"Any"},ticks:{stepSize:1,autoSkip:!1}},y:{title:{display:!0,text:"Mitjana"},suggestedMin:(()=>{const e=i.flatMap(e=>e.data).filter(e=>null!=e),t=e.length?Math.min(...e):0;return Math.floor(10*(t-.1))/10})(),suggestedMax:(()=>{const e=i.flatMap(e=>e.data).filter(e=>null!=e),t=e.length?Math.max(...e):1;return Math.ceil(10*(t+.1))/10})(),ticks:{beginAtZero:!1,stepSize:.1}}},responsive:!0,maintainAspectRatio:!1}}))}function n(){const e=document.getElementById("content");e.innerHTML="";const n=document.createElement("table"),o=document.createElement("tr");["Posició","Jugador","Mitjana"].forEach(e=>{const t=document.createElement("th");t.textContent=e,o.appendChild(t)}),n.appendChild(o);l.filter(e=>parseInt(e.Any,10)===s&&e.Modalitat===d).sort((e,t)=>parseFloat(t.Mitjana)-parseFloat(e.Mitjana)).forEach((e,o)=>{const l=document.createElement("tr");["Posició","Jugador","Mitjana"].forEach(n=>{const a=document.createElement("td");let s;s="Posició"===n?o+1:"Mitjana"===n?Number.parseFloat(e[n]).toFixed(3):"Jugador"===n?e.NomComplet:e[n],a.textContent=s,"Jugador"===n&&(a.classList.add("player-cell"),a.addEventListener("click",n=>{n.stopPropagation(),t(e.Jugador,e.NomComplet)})),l.appendChild(a)}),l.addEventListener("click",()=>{t(e.Jugador,e.NomComplet)}),n.appendChild(l)}),p(e,n)}function o(){const e=document.getElementById("content");e.innerHTML="";const t=document.createElement("table"),n=document.createElement("tr");["#","Jugador","PJ","P","C","E","MG","MM"].forEach(e=>{const t=document.createElement("th");t.textContent=e,n.appendChild(t)}),t.appendChild(n);i.filter(e=>parseInt(e.Any,10)===y&&e.Modalitat===m&&e.Categoria===u).sort((e,t)=>parseInt(e.Posició,10)-parseInt(t.Posició,10)).forEach(e=>{const n=document.createElement("tr");[e["Posició"],e.Jugador||e.Nom||"",e.PartidesJugades||e["Partides jugades"]||e.PJ||"",e.Punts,e.Caramboles,e.Entrades,e.MitjanaGeneral||e.Mitjana||"",e.MitjanaParticular||e["Millor mitjana"]||""].forEach((e,t)=>{const o=document.createElement("td");if(t>=6&&""!==e){const t=Number.parseFloat(e);e=Number.isNaN(t)?e:t.toFixed(3)}o.textContent=e,n.appendChild(o)}),t.appendChild(n)}),p(e,t)}let l=[],a=[],s=null,d="3 BANDES",c=null,i=[],r=[],y=null,m="3 BANDES",u=null;function g(e){c=e}function p(e,t){const n=document.createElement("div");n.className="table-responsive",n.appendChild(t),e.appendChild(n)}function E(){const e=document.getElementById("player-chart");e&&(e.style.width="90vw",e.style.height="96vh");const t=document.getElementById("chart-canvas");t&&(t.style.width="100%",t.style.height="100%"),c&&c.resize()}(()=>{const e=new Date;e.setHours(0,0,0,0),e.setDate(e.getDate()-(e.getDay()+6)%7)})(),"serviceWorker"in navigator&&window.addEventListener("load",()=>{navigator.serviceWorker.register("./service-worker.js").then(e=>{console.log("SW registrat!",e);const t=()=>{e.update().then(()=>{e.waiting&&e.waiting.postMessage({type:"SKIP_WAITING"})})};t(),setInterval(t,864e5)}).catch(e=>console.error("Error al registrar el SW:",e));let e=!1;navigator.serviceWorker.addEventListener("controllerchange",()=>{e||(window.location.reload(),e=!0)})}),document.getElementById("btn-update").addEventListener("click",()=>{"serviceWorker"in navigator&&navigator.serviceWorker.getRegistration().then(e=>{e&&e.update().then(()=>{e.waiting&&e.waiting.postMessage({type:"SKIP_WAITING"})})})}),document.querySelectorAll(".menu-toggle").forEach(e=>{e.addEventListener("click",()=>{const t=document.getElementById(e.dataset.target),n="flex"===t.style.display;if(document.querySelectorAll(".submenu").forEach(e=>{e.style.display="none",e.setAttribute("aria-hidden","true")}),document.querySelectorAll(".menu-toggle").forEach(e=>e.setAttribute("aria-expanded","false")),n)t.style.display="none",t.setAttribute("aria-hidden","true"),e.setAttribute("aria-expanded","false");else{t.style.display="flex",t.setAttribute("aria-hidden","false"),e.setAttribute("aria-expanded","true");const n=t.querySelector("button");n&&n.focus()}document.getElementById("filters-row").style.display="none",document.getElementById("classificacio-filters").style.display="none",document.getElementById("torneig-buttons").style.display="none",document.getElementById("torneig-title").style.display="none",document.getElementById("torneig-category-buttons").style.display="none",document.getElementById("continu3b-buttons").style.display="none",document.getElementById("content").style.display="none"})}),document.getElementById("btn-ranking").addEventListener("click",()=>{document.getElementById("filters-row").style.display="flex",document.getElementById("classificacio-filters").style.display="none",document.getElementById("torneig-buttons").style.display="none",document.getElementById("torneig-title").style.display="none",document.getElementById("torneig-category-buttons").style.display="none",document.getElementById("continu3b-buttons").style.display="none",document.getElementById("content").style.display="block",mostraRanquing()}),document.getElementById("btn-classificacio").addEventListener("click",()=>{document.getElementById("filters-row").style.display="none",document.getElementById("classificacio-filters").style.display="flex",document.getElementById("torneig-buttons").style.display="none",document.getElementById("torneig-title").style.display="none",document.getElementById("torneig-category-buttons").style.display="none",document.getElementById("continu3b-buttons").style.display="none",document.getElementById("content").style.display="block",mostraClassificacio()}),document.getElementById("btn-agenda").addEventListener("click",()=>{document.getElementById("filters-row").style.display="none",document.getElementById("classificacio-filters").style.display="none",document.getElementById("torneig-buttons").style.display="none",document.getElementById("torneig-title").style.display="none",document.getElementById("torneig-category-buttons").style.display="none",document.getElementById("continu3b-buttons").style.display="none",document.getElementById("content").style.display="block",mostraAgenda()}),document.getElementById("btn-horari").addEventListener("click",()=>{document.getElementById("filters-row").style.display="none",document.getElementById("classificacio-filters").style.display="none",document.getElementById("torneig-buttons").style.display="none",document.getElementById("torneig-title").style.display="none",document.getElementById("torneig-category-buttons").style.display="none",document.getElementById("continu3b-buttons").style.display="none",document.getElementById("content").style.display="block",mostraHorari()}),document.getElementById("btn-enllacos").addEventListener("click",()=>{document.getElementById("filters-row").style.display="none",document.getElementById("classificacio-filters").style.display="none",document.getElementById("torneig-buttons").style.display="none",document.getElementById("torneig-title").style.display="none",document.getElementById("torneig-category-buttons").style.display="none",document.getElementById("continu3b-buttons").style.display="none",document.getElementById("content").style.display="block",mostraEnllacos()}),document.getElementById("btn-continu3b").addEventListener("click",()=>{document.getElementById("filters-row").style.display="none",document.getElementById("classificacio-filters").style.display="none",document.getElementById("torneig-buttons").style.display="none",document.getElementById("torneig-title").style.display="none",document.getElementById("torneig-category-buttons").style.display="none";document.getElementById("content").style.display="block";document.getElementById("continu3b-buttons").style.display="flex",mostraContinu3B()}),document.getElementById("btn-torneig").addEventListener("click",()=>{document.getElementById("filters-row").style.display="none",document.getElementById("classificacio-filters").style.display="none",document.getElementById("torneig-buttons").style.display="flex",document.getElementById("torneig-category-buttons").style.display="none",document.getElementById("continu3b-buttons").style.display="none";const e=document.getElementById("content");e.style.display="block",e.innerHTML="",fetch("data/modalitat.json").then(e=>e.json()).then(e=>{!Array.isArray(e)||e[0];const t=document.getElementById("torneig-title");t.textContent=`Social Modalitat ${Array.isArray(e)?e.map(e=>e.Modalitat).join(", "):e.Modalitat||""}`,t.style.display="block"}).catch(()=>{document.getElementById("torneig-title").style.display="none"})}),document.querySelectorAll("#torneig-buttons button").forEach(e=>{e.addEventListener("click",()=>{const t=e.dataset.file;fetch(`data/${t}`).then(e=>e.json()).then(e=>mostraTorneig(e,t)).catch(()=>{document.getElementById("content").innerHTML="<p>Error carregant dades.</p>"})})}),document.getElementById("close-chart").addEventListener("click",function(){document.getElementById("chart-overlay").style.display="none";const t=document.getElementById("chart-title");t&&(t.textContent=""),c&&(c.destroy(),g(null)),e&&(e.focus(),e=null)}),window.addEventListener("resize",function(){const e=document.getElementById("chart-overlay");e&&"none"!==e.style.display&&E()}),Promise.all([fetch("data/ranquing.json").then(e=>e.json()),fetch("data/classificacions.json").then(e=>e.json()).catch(()=>[]),fetch("data/events.json").then(e=>e.json()).catch(()=>[]),fetch("data/calendari.json").then(e=>e.json()).catch(()=>[]),fetch("data/festius.json").then(e=>e.json()).catch(()=>[])]).then(([e,t,c,g,p])=>{l=e,a=[...new Set(e.map(e=>parseInt(e.Any,10)))].sort((e,t)=>e-t),s=a[a.length-1],i=t,r=[...new Set(t.map(e=>parseInt(e.Any,10)))].sort((e,t)=>e-t),y=r[r.length-1]||null;const E=new Set(t.map(e=>e.Categoria));u=E.values().next().value||null,c.map(e=>{const t=e["Títol"]||e.Titol||"";return{...e,"Títol":t,Tipus:t.toLowerCase().includes("assemblea")?"assemblea":"altre"}}).concat(g.map(e=>({Data:e.Data,Hora:"","Títol":`${e["Jugador A"].trim()} vs ${e["Jugador B"].trim()} (${e.Hora})`,Tipus:"partida"}))).concat(p.map(e=>({Data:e.Data,Hora:"","Títol":`Foment tancat (${e.Títol||""})`,Tipus:"festiu"}))),function(){const e=document.getElementById("year-select");e.innerHTML="",a.slice().sort((e,t)=>t-e).forEach(t=>{const n=document.createElement("option");n.value=t,n.textContent=t,t===s&&(n.selected=!0),e.appendChild(n)}),e.addEventListener("change",()=>{s=parseInt(e.value,10),n()});const t=document.getElementById("modalitat-buttons");t.querySelectorAll("button").forEach(e=>{e.dataset.mod===d&&e.classList.add("selected"),e.addEventListener("click",()=>{d=e.dataset.mod,t.querySelectorAll("button").forEach(e=>e.classList.remove("selected")),e.classList.add("selected"),n()})})}(),function(){const e=document.getElementById("classificacio-year-select");e.innerHTML="",r.slice().sort((e,t)=>t-e).forEach(t=>{const n=document.createElement("option");n.value=t,n.textContent=t,t===y&&(n.selected=!0),e.appendChild(n)}),e.addEventListener("change",()=>{y=parseInt(e.value,10),o()});const t=document.getElementById("classificacio-modalitat-buttons");t.querySelectorAll("button").forEach(e=>{e.dataset.mod===m&&e.classList.add("selected"),e.addEventListener("click",()=>{m=e.dataset.mod,t.querySelectorAll("button").forEach(e=>e.classList.remove("selected")),e.classList.add("selected"),o()})});const n=document.getElementById("categoria-buttons");n.innerHTML="",[...new Set(i.map(e=>e.Categoria))].sort().forEach(e=>{const t=document.createElement("button");t.textContent=e,e===u&&t.classList.add("selected"),t.addEventListener("click",()=>{u=e,n.querySelectorAll("button").forEach(e=>e.classList.remove("selected")),t.classList.add("selected"),o()}),n.appendChild(t)})}(),document.getElementById("btn-agenda").click()}).catch(e=>{console.error("Error carregant dades",e)});
+//# sourceMappingURL=main.js.map
