+++ package.json
@@ -1,0 +1,12 @@
+{
+  "name": "foment",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+    "build": "rollup -c"
+  },
+  "devDependencies": {
+    "rollup": "^4.17.0",
+    "@rollup/plugin-terser": "^0.4.4"
+  }
+}
