+++ src/js/partides.js
