+++ src/js/sw-register.js
